requirements.txt
discord.py-self
yt-dlp
aiofiles
dnspython
aiosqlite
psutil